07.phrase07/pyproject.toml
[project]
name = "07-phrase07"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langsmith>=0.7.22",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
]
